Use cellAt() method, un-screw TTT-win detection
tictactoe-win-checking used to be this godawful convolution I don't even want to think about. Among other bad things, it only checked in one direction, which meant it couldn't handle win-checking from the middle of a win: to account for this, I made it seek to the end of a given run before checking whether it was a win. That's just one of the many design oofs it contained. I devised this way-simpler solution as a quick example for a dude's connect-4 game in the AP Students discord, then realized it'd work fine here too